requirements.txt
apache-airflow==2.6.3
finance-datareader==0.9.50
pandas==2.0.3
selenium==4.10.0
beautifulsoup4==4.12.2
numpy
webdriver-manager
pymysql
mysqlclient
pyarrow